requirements.txt
asgiref==3.8.1
certifi==2024.7.4
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gunicorn==23.0.0
itsdangerous==2.2.0
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.9
PyJWT==2.9.0
python-decouple==3.8
setuptools==72.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
whitenoise==6.7.0
